# src/modules/api-docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Store Project 🛍️",
    "description": "A store project that has many routine and professional features of a good store ✅",
    "version": "1.0.0"
  },
...
  "paths": {
    "/auth/register": {
      "get": {
        "tags": ["Auth"],
        "summary": "get register page view"
      },
      "post": {
        "tags": ["Auth"],
        "summary": "user register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "username": "رضا محمدی",
                  "email": "reza@gmail.com",
                  "phone": "09121234567",
                  "password": "rEza123*",
                  "confirmPassword": "rEza123*"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User register successfully"
          },
          "409": {
            "description": "Invalid data"
          }
        }
      }
    },
    "/auth/login": {
      "get": {
        "tags": ["Auth"],
        "summary": "get login page view"
      },
      "post": {
        "tags": ["Auth"],
        "summary": "user login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "phone": "09121234567",
                  "password": "rEza123*"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User login successfully"
          },
          "409": {
            "description": "Invalid data"
          }
        }
      }
    },
...
  }
}